Separate Nexmark static utils into static class
