+++ sdks/java/testing/nexmark/src/main/java/org/apache/beam/sdk/nexmark/queries/Query3.java
@@ -92,18 +92,18 @@ public class Query3 extends NexmarkQuery {
     PCollection<KV<Long, Auction>> auctionsBySellerId =
         eventsWindowed
             // Only want the new auction events.
-            .apply(JUST_NEW_AUCTIONS)
+            .apply(NexmarkQueryUtil.JUST_NEW_AUCTIONS)
 
             // We only want auctions in category 10.
             .apply(name + ".InCategory", Filter.by(auction -> auction.category == 10))
 
             // Key auctions by their seller id.
-            .apply("AuctionBySeller", AUCTION_BY_SELLER);
+            .apply("AuctionBySeller", NexmarkQueryUtil.AUCTION_BY_SELLER);
 
     PCollection<KV<Long, Person>> personsById =
         eventsWindowed
             // Only want the new people events.
-            .apply(JUST_NEW_PERSONS)
+            .apply(NexmarkQueryUtil.JUST_NEW_PERSONS)
 
             // We only want people in OR, ID, CA.
             .apply(
@@ -115,13 +115,13 @@ public class Query3 extends NexmarkQuery {
                             || "CA".equals(person.state)))
 
             // Key people by their id.
-            .apply("PersonById", PERSON_BY_ID);
+            .apply("PersonById", NexmarkQueryUtil.PERSON_BY_ID);
 
     return
     // Join auctions and people.
     // concatenate KeyedPCollections
-    KeyedPCollectionTuple.of(AUCTION_TAG, auctionsBySellerId)
-        .and(PERSON_TAG, personsById)
+    KeyedPCollectionTuple.of(NexmarkQueryUtil.AUCTION_TAG, auctionsBySellerId)
+        .and(NexmarkQueryUtil.PERSON_TAG, personsById)
         // group auctions and persons by personId
         .apply(CoGroupByKey.create())
         .apply(name + ".Join", ParDo.of(joinDoFn))
@@ -213,7 +213,7 @@ public class Query3 extends NexmarkQuery {
         // We've already seen the new person event for this person id.
         // We can join with any new auctions on-the-fly without needing any
         // additional persistent state.
-        for (Auction newAuction : c.element().getValue().getAll(AUCTION_TAG)) {
+        for (Auction newAuction : c.element().getValue().getAll(NexmarkQueryUtil.AUCTION_TAG)) {
           newAuctionCounter.inc();
           newOldOutputCounter.inc();
           c.output(KV.of(newAuction, existingPerson));
@@ -222,7 +222,7 @@ public class Query3 extends NexmarkQuery {
       }
 
       Person theNewPerson = null;
-      for (Person newPerson : c.element().getValue().getAll(PERSON_TAG)) {
+      for (Person newPerson : c.element().getValue().getAll(NexmarkQueryUtil.PERSON_TAG)) {
         if (theNewPerson == null) {
           theNewPerson = newPerson;
         } else {
@@ -246,7 +246,7 @@ public class Query3 extends NexmarkQuery {
           auctionsState.clear();
         }
         // Also deal with any new auctions.
-        for (Auction newAuction : c.element().getValue().getAll(AUCTION_TAG)) {
+        for (Auction newAuction : c.element().getValue().getAll(NexmarkQueryUtil.AUCTION_TAG)) {
           newAuctionCounter.inc();
           newNewOutputCounter.inc();
           c.output(KV.of(newAuction, newPerson));
@@ -268,7 +268,7 @@ public class Query3 extends NexmarkQuery {
       if (pendingAuctions == null) {
         pendingAuctions = new ArrayList<>();
       }
-      for (Auction newAuction : c.element().getValue().getAll(AUCTION_TAG)) {
+      for (Auction newAuction : c.element().getValue().getAll(NexmarkQueryUtil.AUCTION_TAG)) {
         newAuctionCounter.inc();
         pendingAuctions.add(newAuction);
       }
