+++ sdks/java/testing/nexmark/src/main/java/org/apache/beam/sdk/nexmark/queries/sql/SqlQuery1.java
@@ -17,13 +17,12 @@
  */
 package org.apache.beam.sdk.nexmark.queries.sql;
 
-import static org.apache.beam.sdk.nexmark.queries.NexmarkQuery.IS_BID;
-
 import org.apache.beam.sdk.extensions.sql.SqlTransform;
 import org.apache.beam.sdk.nexmark.model.Bid;
 import org.apache.beam.sdk.nexmark.model.Event;
 import org.apache.beam.sdk.nexmark.model.Event.Type;
 import org.apache.beam.sdk.nexmark.model.sql.SelectEvent;
+import org.apache.beam.sdk.nexmark.queries.NexmarkQueryUtil;
 import org.apache.beam.sdk.schemas.transforms.Convert;
 import org.apache.beam.sdk.transforms.Filter;
 import org.apache.beam.sdk.transforms.PTransform;
@@ -65,7 +64,7 @@ public class SqlQuery1 extends PTransform<PCollection<Event>, PCollection<Bid>> 
   @Override
   public PCollection<Bid> expand(PCollection<Event> allEvents) {
     return allEvents
-        .apply(Filter.by(IS_BID))
+        .apply(Filter.by(NexmarkQueryUtil.IS_BID))
         .apply(getName() + ".SelectEvent", new SelectEvent(Type.BID))
         .apply(QUERY)
         .apply(Convert.fromRows(Bid.class));
